requirements.txt
mysql.connector
flask
os
dotenv